+++ com.unity.render-pipelines.high-definition/Editor/Material/Lit/ShaderGraph/HDLitSettingsView.cs
@@ -42,9 +42,45 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline.Drawing
                 });
             });
 
+            ++indentLevel;
+            ps.Add(new PropertyRow(CreateLabel("Rendering Pass", indentLevel)), (row) =>
+            {
+                Enum defaultValue;
+                switch (m_Node.surfaceType)
+                {
+                    case SurfaceType.Opaque:
+                        defaultValue = HDRenderQueue.OpaqueRenderQueue.Default;
+                        break;
+                    case SurfaceType.Transparent:
+                        defaultValue = HDRenderQueue.TransparentRenderQueue.Default;
+                        break;
+                    default:
+                        throw new ArgumentException("Unknown SurfaceType");
+                }
+                row.Add(new EnumField(defaultValue), (field) =>
+                {
+                    switch (m_Node.surfaceType)
+                    {
+                        case SurfaceType.Opaque:
+                            //GetOpaqueEquivalent: prevent issue when switching surface type
+                            field.value = HDRenderQueue.ConvertToOpaqueRenderQueue(HDRenderQueue.GetOpaqueEquivalent(m_Node.renderingPass));
+                            break;
+                        case SurfaceType.Transparent:
+                            //GetTransparentEquivalent: prevent issue when switching surface type
+                            field.value = HDRenderQueue.ConvertToTransparentRenderQueue(HDRenderQueue.GetTransparentEquivalent(m_Node.renderingPass));
+                            break;
+                        default:
+                            throw new ArgumentException("Unknown SurfaceType");
+                    }
+                    field.RegisterValueChangedCallback(ChangeRenderingPass);
+                });
+            });
+            --indentLevel;
+
             if (m_Node.surfaceType == SurfaceType.Transparent)
             {
                 ++indentLevel;
+
                 if (!m_Node.HasRefraction())
                 {
                     ps.Add(new PropertyRow(CreateLabel("Blending Mode", indentLevel)), (row) =>
@@ -322,6 +358,27 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline.Drawing
             m_Node.owner.owner.RegisterCompleteObjectUndo("Alpha Mode Change");
             m_Node.alphaMode = alphaMode;
         }
+        void ChangeRenderingPass(ChangeEvent<Enum> evt)
+        {
+            HDRenderQueue.RenderQueueType renderingPass;
+            switch (m_Node.surfaceType)
+            {
+                case SurfaceType.Opaque:
+                    renderingPass = HDRenderQueue.ConvertFromOpaqueRenderQueue((HDRenderQueue.OpaqueRenderQueue)evt.newValue);
+                    break;
+                case SurfaceType.Transparent:
+                    renderingPass = HDRenderQueue.ConvertFromTransparentRenderQueue((HDRenderQueue.TransparentRenderQueue)evt.newValue);
+                    break;
+                default:
+                    throw new ArgumentException("Unknown SurfaceType");
+            }
+
+            if (Equals(m_Node.renderingPass, renderingPass))
+                return;
+
+            m_Node.owner.owner.RegisterCompleteObjectUndo("Rendering Pass Change");
+            m_Node.renderingPass = renderingPass;
+        }
 
         void ChangeBlendPreserveSpecular(ChangeEvent<bool> evt)
         {
@@ -337,14 +394,6 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline.Drawing
             ToggleData td = m_Node.transparencyFog;
             td.isOn = evt.newValue;
             m_Node.transparencyFog = td;
-        }
-
-        void ChangeDrawBeforeRefraction(ChangeEvent<bool> evt)
-        {
-            m_Node.owner.owner.RegisterCompleteObjectUndo("Draw Before Refraction Change");
-            ToggleData td = m_Node.drawBeforeRefraction;
-            td.isOn = evt.newValue;
-            m_Node.drawBeforeRefraction = td;
         }
 
         void ChangeRefractionModel(ChangeEvent<Enum> evt)
@@ -391,7 +440,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline.Drawing
 
         void ChangeSortPriority(ChangeEvent<int> evt)
         {
-            m_Node.sortPriority = Math.Max(-HDRenderQueue.k_TransparentPriorityQueueRange, Math.Min(evt.newValue, HDRenderQueue.k_TransparentPriorityQueueRange));
+            m_Node.sortPriority = HDRenderQueue.ClampsTransparentRangePriority(evt.newValue);
             // Force the text to match.
             m_SortPiorityField.value = m_Node.sortPriority;
             if (Equals(m_Node.sortPriority, evt.newValue))