+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Wizard helping to configure HDRP project.
 - Added new UI for decal material to allow remapping and scaling of some properties
 - Added cascade shadow visualisation toggle in HD shadow settings
+- Added new Rendering Pass in materials.
 - Added icons for assets.
 - Added replace blending mode for distortion
 - Added post-processing features (work in progress)